requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.34.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
pdf2image==1.16.3
pytesseract==0.3.10